requirements.txt
# Core
torch>=2.2
transformers>=4.44
datasets>=3.0
accelerate>=1.0
evaluate>=0.4
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
PyYAML>=6.0
tqdm>=4.66

# Optional utils
python-dotenv>=1.0
rich>=13.7
tyro>=0.8

# LLM APIs
google-generativeai>=0.8.0
openai>=2.0.0,<3.0.0
anthropic>=0.30.0
langchain-openai>=0.3.0

# Agent Frameworks
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.40
langsmith>=0.0.80


